backend/requirements.txt
Flask
Flask-CORS
requests
PyJWT
